Move parse_test -> vm_test
