+++ vms/saevm/cchain/vm_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/ava-labs/libevm/core/types"
 	"github.com/ava-labs/libevm/ethclient"
 	"github.com/ava-labs/libevm/libevm/options"
+	"github.com/ava-labs/libevm/rlp"
 	"github.com/google/go-cmp/cmp"
 	"github.com/holiman/uint256"
 	"github.com/stretchr/testify/assert"
@@ -653,4 +654,84 @@ func TestDebugTraceDoesNotApplyAtomicState(t *testing.T) {
 	// execution results to have been applied.
 	exportedUTXOs := txtest.ExportedUTXOs(signedExport.ID(), export)
 	sut.assertUTXOsMissing(t, destinationChain, sut.ctx.ChainID, exportedUTXOs...)
+}
+
+// newTamperedBlock returns a block that encodes txs but whose header commits an
+// ExtDataHash that does not match its extData, simulating tampering.
+func newTamperedBlock(tb testing.TB, number uint64, parent common.Hash, txs ...*tx.Tx) *types.Block {
+	tb.Helper()
+
+	extData, err := tx.MarshalSlice(txs)
+	require.NoErrorf(tb, err, "tx.MarshalSlice(%d txs)", len(txs))
+
+	header := customtypes.WithHeaderExtra(
+		&types.Header{
+			ParentHash: parent,
+			Number:     new(big.Int).SetUint64(number),
+		},
+		&customtypes.HeaderExtra{ExtDataHash: common.Hash(ids.GenerateTestID())},
+	)
+	return customtypes.NewBlockWithExtData(
+		header,
+		nil, // txs
+		nil, // uncles
+		nil, // receipts
+		saetest.TrieHasher(),
+		extData,
+		false, // keep the mismatching ExtDataHash; do not recompute
+	)
+}
+
+// TestParseBlock verifies that the cchain ParseBlock override accepts
+// well-formed blocks and rejects blocks whose extData does not match the
+// ExtDataHash committed in the header.
+func TestParseBlock(t *testing.T) {
+	ctx, sut := newSUT(t)
+
+	w := newWallet(txtest.NewKey(t), snowtest.Context(t, snowtest.CChainID), nil)
+	tx1 := w.newMinimalTx(t)
+
+	tests := []struct {
+		name    string
+		block   *types.Block
+		wantErr error
+		wantTxs []ids.ID
+	}{
+		{
+			name:    "valid",
+			block:   newBlock(t, 1, common.Hash{}, tx1),
+			wantTxs: []ids.ID{tx1.ID()},
+		},
+		{
+			name:    "valid_empty",
+			block:   newBlock(t, 1, common.Hash{}),
+			wantTxs: []ids.ID{},
+		},
+		{
+			name:    "extdata_hash_mismatch",
+			block:   newTamperedBlock(t, 1, common.Hash{}, tx1),
+			wantErr: errExtDataHashMismatch,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			buf, err := rlp.EncodeToBytes(tt.block)
+			require.NoError(t, err, "rlp.EncodeToBytes(block)")
+
+			got, err := sut.ParseBlock(ctx, buf)
+			require.ErrorIs(t, err, tt.wantErr, "vm.ParseBlock(buf)")
+			if tt.wantErr != nil {
+				return
+			}
+
+			gotTxs, err := tx.ParseSlice(customtypes.BlockExtData(got.EthBlock()))
+			require.NoError(t, err, "tx.ParseSlice(extData)")
+
+			gotIDs := make([]ids.ID, len(gotTxs))
+			for i, txn := range gotTxs {
+				gotIDs[i] = txn.ID()
+			}
+			require.Equal(t, tt.wantTxs, gotIDs, "parsed cross-chain tx IDs")
+		})
+	}
 }
